refactor: ExtractTextFromMessage -> chathistories
