+++ internal/models/chathistories/chat_histories.go
@@ -4,16 +4,20 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"net/url"
 	"strconv"
 	"strings"
 	"text/template"
 	"time"
+	"unicode/utf16"
 	"unicode/utf8"
 
 	"entgo.io/ent/dialect/sql"
+	"github.com/Junzki/link-preview"
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/google/uuid"
 	"github.com/samber/lo"
+	lop "github.com/samber/lo/parallel"
 	"github.com/sirupsen/logrus"
 	"go.uber.org/fx"
 
@@ -50,8 +54,60 @@ func NewModel() func(NewModelParams) (*Model, error) {
 	}
 }
 
+func (m *Model) ExtractTextFromMessage(message *tgbotapi.Message) string {
+	text := lo.Ternary(message.Caption != "", message.Caption, message.Text)
+
+	type MarkdownLink struct {
+		Markdown []uint16
+		Start    int
+		End      int
+	}
+
+	textUTF16 := utf16.Encode([]rune(text))
+	links := lop.Map(message.Entities, func(entity tgbotapi.MessageEntity, i int) MarkdownLink {
+		startIndex := entity.Offset
+		endIndex := startIndex + entity.Length
+		var title string
+		var href string
+		if entity.Type == "url" {
+			href = string(utf16.Decode(textUTF16[startIndex:endIndex]))
+			result, err := LinkPreview.PreviewLink(href, nil)
+			if err != nil {
+				m.logger.Infof("🔗Failed to generate link preview for %s, error %+v", href, err)
+				return MarkdownLink{[]uint16{}, -1, -1}
+			}
+			title = result.Title
+		} else if entity.Type == "text_link" {
+			title = string(utf16.Decode(textUTF16[startIndex:endIndex]))
+			href = entity.URL
+		} else {
+			return MarkdownLink{[]uint16{}, -1, -1}
+		}
+
+		unescaped, err := url.QueryUnescape(href)
+		if err != nil {
+			href = strings.ReplaceAll(unescaped, " ", "+")
+		}
+
+		md := "[" + title + "](" + href + ")"
+		mdUTF16 := utf16.Encode([]rune(md))
+
+		return MarkdownLink{mdUTF16, startIndex, endIndex}
+	})
+
+	for i := len(links) - 1; i >= 0; i-- {
+		if links[i].Start == -1 {
+			continue
+		}
+		temp := append(links[i].Markdown, textUTF16[links[i].End:]...)
+		textUTF16 = append(textUTF16[:links[i].Start], temp...)
+	}
+
+	return string(utf16.Decode(textUTF16))
+}
+
 func (m *Model) extractTextWithSummarization(message *tgbotapi.Message) (string, error) {
-	text := tgbot.ExtractTextFromMessage(message)
+	text := m.ExtractTextFromMessage(message)
 	if text == "" {
 		return "", nil
 	}
