+++ pkg/bots/tgbot/utils.go
@@ -5,12 +5,6 @@ import (
 	"net/url"
 	"regexp"
 	"strings"
-	"unicode/utf16"
-
-	"github.com/Junzki/link-preview"
-	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
-	"github.com/samber/lo"
-	lop "github.com/samber/lo/parallel"
 
 	"github.com/nekomeowww/insights-bot/pkg/types/telegram"
 	"github.com/nekomeowww/insights-bot/pkg/utils"
@@ -83,54 +77,6 @@ func FullNameFromFirstAndLastName(firstName, lastName string) string {
 	}
 
 	return firstName + " " + lastName
-}
-
-func ExtractTextFromMessage(message *tgbotapi.Message) string {
-	text := lo.Ternary(message.Caption != "", message.Caption, message.Text)
-
-	type MarkdownLink struct {
-		Markdown []uint16
-		Start    int
-		End      int
-	}
-
-	textUtf16 := utf16.Encode([]rune(text))
-	links := lop.Map(message.Entities, func(entity tgbotapi.MessageEntity, i int) MarkdownLink {
-		startIndex := entity.Offset
-		endIndex := startIndex + entity.Length
-		var title string
-		var href string
-		if entity.Type == "url" {
-			href = string(utf16.Decode(textUtf16[startIndex:endIndex]))
-			result, err := LinkPreview.PreviewLink(href, nil)
-			if err != nil {
-				return MarkdownLink{[]uint16{}, -1, -1}
-			}
-			title = result.Title
-		} else if entity.Type == "text_link" {
-			title = string(utf16.Decode(textUtf16[startIndex:endIndex]))
-			href = entity.URL
-		} else {
-			return MarkdownLink{[]uint16{}, -1, -1}
-		}
-		unescaped, err := url.QueryUnescape(href)
-		if err == nil {
-			href = strings.ReplaceAll(unescaped, " ", "+")
-		}
-		md := "[" + title + "](" + href + ")"
-		mdUtf16 := utf16.Encode([]rune(md))
-		return MarkdownLink{mdUtf16, startIndex, endIndex}
-	})
-
-	for i := len(links) - 1; i >= 0; i-- {
-		if links[i].Start == -1 {
-			continue
-		}
-		temp := append(links[i].Markdown, textUtf16[links[i].End:]...)
-		textUtf16 = append(textUtf16[:links[i].Start], temp...)
-	}
-
-	return string(utf16.Decode(textUtf16))
 }
 
 // EscapeHTMLSymbols
